+++ apps/server/src/shared/domain/entity/submission.entity.spec.ts
@@ -1,6 +1,5 @@
 import { Collection } from '@mikro-orm/core';
-import { Test, TestingModule } from '@nestjs/testing';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { setupEntities } from '@src/modules/database';
 import { userFactory } from '../factory';
 import { File } from './file.entity';
 import { Submission } from './submission.entity';
@@ -14,14 +13,8 @@ const buildSubmission = () => {
 };
 
 describe('Submission entity', () => {
-	let module: TestingModule;
-
 	beforeAll(async () => {
-		module = await Test.createTestingModule({ imports: [MongoMemoryDatabaseModule.forRoot()] }).compile();
-	});
-
-	afterAll(async () => {
-		await module.close();
+		await setupEntities();
 	});
 
 	it('should be graded if grade percentage is set', () => {