+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
+- BC-487 - update mongodb-memory-server from 6.9.6 to 7.4.4
+- BC-490 - setup entities for tests without needing a database
+
 ## [26.12.0] - 2021-10-12
 
 ### Added